add: gencode of struct

+++ struct-inputs/input-2.rs
@@ -1,10 +1,10 @@
 struct Point {
-    x: i64,
-    y: i64,
+    x: i32,
+    y: i32,
 }
 
 static ORIGIN: Point = Point { x: 0, y: 0 };
 
-fn main() {
-    return(0)
+fn main(){
+    return (0)
 }
